paper_research_system/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
aiohttp>=3.8.0
networkx>=3.0.0
numpy>=1.21.0
scipy>=1.7.0